Resolve test failure by scoping implicits.
We need the int instance to ensure increment |+| works as you would
expect.

We need the BoundedSemiLattice instance to handle the max behaviour via
a |+| b. Unfortuantely since both are Monoid[Int] we cannot have both in
scope at the same time.

For now I've scoped the required implicits explicitly in the tests to
ensure the correct implicit is available for the behaviour that we need.